[FIX] l10n_br_repair: remove methods from fiscal_line_mixin.py and put on repair_fee.py and repair_line.py
